+++ src/api/reddit/models/download_meta.rs
@@ -5,8 +5,8 @@ use crate::api::config::configuration::Configuration;
 pub struct DownloadMeta {
 	pub url: String,
 	pub subreddit_name: String,
-	pub image_height: u32,
-	pub image_width: u32,
+	pub image_height: usize,
+	pub image_width: usize,
 	pub post_link: String,
 	pub nsfw: bool,
 	pub filename: String,
